% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de}
\alias{de}
\title{Differential expression results for yeast RNA-seq.}
\format{
A tibble with 6 columns
}
\usage{
data(de)
}
\value{
Results for differential expression for yeast RNA-seq.
}
\description{
A subset of 6 + 6 replicates was selected from data set reported in
https://doi.org/10.1093/bioinformatics/btv425
}
\keyword{datasets}
