% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{dexdash_list}
\alias{dexdash_list}
\title{Create a list of *dexdash* data sets.}
\usage{
dexdash_list(...)
}
\arguments{
\item{...}{One or more objects of class `dexdash_set`.}
}
\value{
A named list of `dexdash_set` objects with class attribute
  `dexdash_list`.
}
\description{
Merge multiple \code{dexdash_set} objects into a \code{dexdah_list}. This
function is used to merge multiple data sets into one object, that can be fed
into the Shiny app. This allows for browsing results from, e.g., different
experiments in one app.
}
