% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{dexdash_set}
\alias{dexdash_set}
\title{Create *dexdash* data set}
\usage{
dexdash_set(de, data, metadata, name)
}
\arguments{
\item{de}{Differential expression data as a data frame, expected to contain
the following columns: `id` - feature id, `log_fc` - log-fold change,
`expr` - expression or abundance, e.g. gene read count, `p_value` -
uncorrected p-value from the DE test, `contrast` - name of the contrast
used for the test.}

\item{data}{A data frame containing the primary dataset for exploration,
expected to contain the following columns: `id` - feature id, the same as
in `de` data frame, `sample` - sample name, the same as in the `metadata`
data frame, `value` - the expression or abundance.}

\item{metadata}{Metadata associated with the `data` parameter, providing
additional context or grouping for the samples or features included in
`data`. Expected to contain the following columns: `sample` - must match
samples in the `data` object, `group` - a grouping variable, e.g.,
condition or treatment, `replicate` - replicate name.}

\item{name}{A string to identify the name of the set.}
}
\value{
An object of class \code{dexdash_set} required by the Shiny app
  launcher \code{run_app}.
}
\description{
This function creates a data set for *dexdash* Shiny app, based on
user-provided data.
}
\examples{
data(yeast_de, yeast_data, yeast_metadata)
dexset <- dexdash_set(yeast_de, yeast_data, yeast_metadata, "Yeast")
}
