% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{download_feature_information}
\alias{download_feature_information}
\title{Download feature (gene or protein) information for a given species}
\usage{
download_feature_information(
  species,
  species_file = NULL,
  id = "ensembl_gene_id"
)
}
\arguments{
\item{species}{A character string specifying the species for which to
download data. Available species can be listed using function
`list_species()`.}

\item{species_file}{(Optional) A character string providing the path to a
JSON file containing species information. If `NULL`, the default
species.json file from the package will be used. See Details for more
information.}

\item{id}{A string with the BioMart attribute used as identifier. Default
value is "ensembl_gene_id", referring to Ensembl gene ID. For UniProt
identifiers, use "uniprotswissprot"; for NCBI identifiers use
"entrezgene_id".}
}
\value{
A tibble containing gene IDs, gene names (cleaned), and gene
  descriptions (cleaned).
}
\description{
Connects to the Ensembl BioMart for the specified species and downloads
feature information including feature IDs, gene symbols, and descriptions. It
cleans up the gene descriptions and handles missing gene names and duplicate
entries.
}
\details{
BioMart requires a biomart name, a host name and a dataset, to
  connect to the right database. For example, for human, these arguments are
  "ensembl", "https://www.ensembl.org", and "hsapiens_gene_ensembl",
  respectively. This package contains a small JSON file
  (can be found at \code{system.file("extdata", "species.json", package =
  "dexdash")}), with Ensembl information for a few species. If
  your species is not included, you need to create a JSON file in the same
  format, as the included file.
}
\examples{
\dontrun{
gene_info <- download_gene_information("mouse")
}
}
