% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{download_functional_terms}
\alias{download_functional_terms}
\title{Download Functional Annotation Terms}
\usage{
download_functional_terms(species, species_file = NULL)
}
\arguments{
\item{species}{A character string specifying the species for which to
download data. Available species can be listed using function
`list_species()`.}

\item{species_file}{(Optional) A character string providing the path to a
JSON file containing species information. If `NULL`, the default
species.json file from the package will be used. See Details for more
information.}
}
\value{
A list of three elements named "go", "reactome" and "kegg", each
  containing two data frames with term descriptions and feature mapping.
}
\description{
Retrieves Gene Ontology (GO), Reactome, and KEGG terms for a given species.
}
\details{
The GO, KEGG and Reactome databases use different species
  designation names. For example, designation for yeast is "YEAST-mod",
  "Saccharomyces cerevisiae" and "sce", for GO, Reactome and KEGG,
  respectively. In oder to interrogate these databases, the correct
  designations must be passed on. This package contains a small JSON file
  (can be found at \code{system.file("extdata", "species.json", package =
  "dexdash")}), with designation information for a few species. If
  your species is not included, you need to create a JSON file in the same
  format, as the included file. The species designations can be found using
  `fenr::fetch_go_species()`, `fenr::fetch_reactome_species()` and
  `fenr::fetch_kegg_species()`. These three functions return data frames,
  where column `designation` contains the species designation required.
}
\examples{
\dontrun{
fterms <- download_functional_terms(species = "yeast", feature_name = "gene_id")
}
}
