% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_link}
\alias{download_link}
\title{Create a Download Link}
\usage{
download_link(id)
}
\arguments{
\item{id}{A character string specifying the ID for the download link.}
}
\value{
A Shiny download link UI component.
}
\description{
This function creates a Shiny download link with a specified ID.
}
