% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features}
\alias{features}
\title{Feature description for an example RNA-seq experiment.}
\format{
A tibble with 6 columns
}
\usage{
data(features)
}
\value{
Feature description table.
}
\description{
Feature description for an example RNA-seq experiment.
}
\keyword{datasets}
