% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gear_icon}
\alias{gear_icon}
\title{Create a Gear Icon with Popover}
\usage{
gear_icon(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the popover content.}
}
\value{
A Shiny UI component with a gear icon and a popover.
}
\description{
This function creates a gear icon that, when clicked, displays a popover with the specified content.
}
