% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{info_icon}
\alias{info_icon}
\title{Create an Info Icon with Popover}
\usage{
info_icon(helper)
}
\arguments{
\item{helper}{A character string specifying the name of the markdown file
(without the extension) located in the "inst/helpers" directory.}
}
\value{
A Shiny UI component with an info icon and a popover.
}
\description{
This function creates an info icon that, when clicked, displays a popover
with content from a markdown file.
}
