% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_genes}
\alias{mouse_genes}
\title{Feature description for an example RNA-seq experiment.}
\format{
A tibble with 3 columns
}
\usage{
data(mouse_genes)
}
\value{
Feature description table, downloaded from Ensembl.
}
\description{
Feature description for an example RNA-seq experiment.
}
\keyword{datasets}
