% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_mtx}
\alias{mouse_mtx}
\title{Expression data for mouse RNA-seq in wide matrix format.}
\format{
A matrix with 50892 rows and 16 columns.
}
\usage{
data(mouse_mtx)
}
\value{
Expression data from a mouse RNA-seq experiment.
}
\description{
Expression data for mouse RNA-seq in wide matrix format.
}
\keyword{datasets}
