% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_terms}
\alias{mouse_terms}
\title{Functional terms for mouse}
\format{
A list with data frames
}
\usage{
data(mouse_terms)
}
\value{
Functional terms for mouse.
}
\description{
Functional terms for mouse
}
\keyword{datasets}
