% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtx}
\alias{mtx}
\title{Expression data for yeast RNA-seq in wide matrix format.}
\format{
A matrix with 6298 rows and 12 columns.
}
\usage{
data(mtx)
}
\value{
Expression data from a yeast RNA-seq experiment.
}
\description{
A subset of 6 + 6 replicates was selected from data set reported in
https://doi.org/10.1093/bioinformatics/btv425
}
\keyword{datasets}
