% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{prepare_functional_terms}
\alias{prepare_functional_terms}
\title{Prepare functional terms for fast enrichment}
\usage{
prepare_functional_terms(
  terms,
  feature_name = c("gene_symbol", "gene_id"),
  all_features = NULL
)
}
\arguments{
\item{terms}{A list of three elements, named "go", "reactome" and "kegg".
Each element is a list with two data frames, "terms" and "mapping". See
\code{vignette("fenr")} for more details. In a normal workflow, this object
is created with \code{download_functional_terms}.}

\item{feature_name}{The name of the column in the \code{mapping} tibble to be
used as the feature identifier. It can be "gene_symbol" or "gene_id". If
your data contain gene symbols (e.g. "BRCA1" or "FOXP1"), use \code{feature_name
= "gene_symbol"}. If your data contain other identifiers (e.g. "ENSG00000012048"
 or  "ENSG00000114861"), use \code{feature_name
= "gene_id"}.}

\item{all_features}{(Optional) A vector of all possible features (such as
gene symbols) to prepare the data for enrichment analysis.}
}
\value{
An object to be used by the Shiny app.
}
\description{
Prepare functional terms for fast enrichment
}
\examples{
\dontrun{
terms <- download_functional_terms(species = "yeast")
fterms <- prepare_functional_terms(terms)
}
}
