% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{run_app}
\alias{run_app}
\title{Launches an interactive differential expression (DE) data explorer}
\usage{
run_app(
  dexset,
  features,
  fterms,
  title = "DE explorer",
  x_variable = "sample",
  colour_variable = NULL
)
}
\arguments{
\item{dexset}{Either a \code{dexdash_set} object containing data from one
set, or a \code{dexdash_list} object containing data from multiple sets.
The former one is created using \code{dexdash_set()} function, the latter
one is created with \code{dexdash_list()} function.}

\item{features}{A data frame that maps feature identifiers to names and
descriptions. Expected to contain the following columns: `id` - feature id,
must match the identifier in the `data` object, `name` - human-friendly
name of the feature, e.g. gene symbol, `description` - a brief description
of the feature. This data frame can be obtained using function
\code{download_gene_informarion()}.}

\item{fterms}{An object containing functional term information, created using
function \code{download_functional_terms}.}

\item{title}{A string with a short title, which is presented at the top of
the side bar.}

\item{x_variable}{A string with the name of the startup x-axis variable for the
feature plot. It should correspond to a column name in the metadata, included
in the dexset object. The default is "sample".}

\item{colour_variable}{A string with the name of the startup colour variable for the
feature plot. It should correspond to a column name in the metadata, included
in the dexset object. If left NULL (default), the second column from metadata
will be used.}
}
\value{
The function does not return a value but launches a Shiny application
  in the user's default web browser, allowing for interactive exploration of
  the differential expression data.
}
\description{
This function creates and launches a Shiny application designed for exploring
differential expression (DE) data. It integrates various data inputs and
initializes interactive visualization modules for an enhanced data
exploration experience. The application offers a sidebar layout with themed
UI components and a range of interactive modules including global input,
volume-magnitude plot, feature plot, feature information, and enrichment
analysis.
}
\examples{
if(interactive()) {
  library(dexdata)
  data(de, data, metadata, features)
  dexset <- dexdash_set(de, data, metadata, "Yeast")
  fterms <- download_functional_terms("yeast", feature_name = "gene_id")
  run_app(dexset, features, fterms)
}
}
