% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_data}
\alias{yeast_data}
\title{Expression data for yeast RNA-seq.}
\format{
A tibble with 3 columns
}
\usage{
data(yeast_data)
}
\value{
Expression data from a yeast RNA-seq experiment.
}
\description{
A subset of 6 + 6 replicates was selected from data set reported in
https://doi.org/10.1093/bioinformatics/btv425
}
\keyword{datasets}
