% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_fterms}
\alias{yeast_fterms}
\title{Functional terms for yeast.}
\format{
A 'fenr_terms' object
}
\usage{
data(mouse_fterms)
}
\value{
Functional terms for yeast.
}
\description{
Functional terms for yeast.
}
\keyword{datasets}
