% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_metadata}
\alias{yeast_metadata}
\title{Experimental design from an yeast RNA-seq experiment.}
\format{
A tibble with 3 columns
}
\usage{
data(yeast_metadata)
}
\value{
Design of the experiment containing expression `data` and
  differential expression `de`.
}
\description{
A subset of 6 + 6 replicates was selected from data set reported in
https://doi.org/10.1093/bioinformatics/btv425
}
\keyword{datasets}
