% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{download_functional_terms}
\alias{download_functional_terms}
\title{Download Functional Annotation Terms}
\usage{
download_functional_terms(
  species,
  species_file = NULL,
  feature_name = c("gene_symbol", "gene_id"),
  all_features = NULL
)
}
\arguments{
\item{species}{A character string specifying the species for which to
download data.}

\item{species_file}{(Optional) A character string providing the path to a
JSON file containing species information. If `NULL`, the default
species.json file from the package will be used.}

\item{feature_name}{The name of the column in the \code{mapping} tibble to be
used as the feature identifier. It can be "gene_symbol" or "gene_id". If
your data contain gene symbols (e.g. "BRCA1" or "FOXP1"), use \code{feature_name
= "gene_symbol"}. If your data contain Ensembl identifiers (e.g. "ENSG00000012048"
 or  "ENSG00000114861"), use \code{feature_name
= "gene_id"}.}

\item{all_features}{(Optional) A vector of all possible features (such as
gene symbols) to prepare the data for enrichment analysis.}
}
\value{
A list of three elements, each containing data frames with the terms
  and mappings for GO, Reactome, and KEGG data respectively. The list
  elements are named "go", "reactome", and "kegg".
}
\description{
Retrieves Gene Ontology (GO), Reactome, and KEGG terms for a given species
and prepares them to use with the Shiny app.
}
\details{
This function checks for the existence of a specified species file
  and verifies that the provided species is present within it. It then
  proceeds to load the term data using `fenr` functions for the different
  biological databases. If `all_features` is provided, it also prepares the
  data for enrichment analysis with `fenr::prepare_for_enrichment`.
}
\examples{
\dontrun{
terms <- download_functional_terms(species = "yeast", feature_name = "gene_id")
}
}
