% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{download_functional_terms}
\alias{download_functional_terms}
\title{Download Functional Annotation Terms}
\usage{
download_functional_terms(
  species,
  species_file = NULL,
  feature_name = c("gene_symbol", "gene_id"),
  all_features = NULL
)
}
\arguments{
\item{species}{A character string specifying the species for which to
download data.}

\item{species_file}{(Optional) A character string providing the path to a
JSON file containing species information. If `NULL`, the default
species.json file from the package will be used. See Details for more
information.}

\item{feature_name}{The name of the column in the \code{mapping} tibble to be
used as the feature identifier. It can be "gene_symbol" or "gene_id". If
your data contain gene symbols (e.g. "BRCA1" or "FOXP1"), use \code{feature_name
= "gene_symbol"}. If your data contain Ensembl identifiers (e.g. "ENSG00000012048"
 or  "ENSG00000114861"), use \code{feature_name
= "gene_id"}.}

\item{all_features}{(Optional) A vector of all possible features (such as
gene symbols) to prepare the data for enrichment analysis.}
}
\value{
A list of three elements named "go", "reactome" and "kegg", each
  containing \code{fenr_terms} objects required for fast functional
  enrichment.
}
\description{
Retrieves Gene Ontology (GO), Reactome, and KEGG terms for a given species
and prepares them to use with the Shiny app.
}
\details{
The GO, KEGG and Reactome databases use different species
  designation names. For example, designation for yeast is "sgd",
  "Saccharomyces cerevisiae" and "sce", for GO, Reactome and KEGG,
  respectively. In oder to interrogate these databases, the correct
  designations must be passed on. This package contains a small JSON file
  (can be found at \code{system.file("extdata", "species.json", package =
  "dexdash")}), with designation information for a few species. If
  your species is not included, you need to create a JSON file in the same
  format, as the included file. The species designations can be found using
  `fenr::fetch_go_species()`, `fenr::fetch_reactome_species()` and
  `fenr::fetch_kegg_species()`. These three functions return data frames,
  where column `designation` contains the species designation required.
}
\examples{
fterms <- download_functional_terms(species = "yeast", feature_name = "gene_id")
}
