% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_mtx}
\alias{mouse_mtx}
\title{Expression data for mouse RNA-seq in wide matrix format.}
\format{
A matrix with 29134 rows and 12 columns.
}
\usage{
data(mouse_mtx)
}
\value{
Expression data from a mouse RNA-seq experiment.
}
\description{
A subset of 3 time points from data reported in
https://doi.org/10.1016/j.celrep.2021.109943
}
\keyword{datasets}
