% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{run_app}
\alias{run_app}
\title{Launches an Interactive Differential Expression (DE) Data Explorer}
\usage{
run_app(de, data, metadata, features, fterms)
}
\arguments{
\item{de}{Differential expression data as a data frame, expected to contain
the following columns: `id` - feature id, `log_fc` - log-fold change,
`expr` - expression or abundance, e.g. gene read count, `p_value` - p-value
from the DE test, `fdr` - false discovery rate, which is p-value corrected
for multiple tests, `contrast` - name of the contrast used for the test.}

\item{data}{A data frame containing the primary dataset for exploration,
expected to contain the following columns: `id` - feature id, the same as
in `de` data frame, `sample` - sample name, the same as in the `metadata`
data frame, `value` - the expression or abundance.}

\item{metadata}{Metadata associated with the `data` parameter, providing
additional context or grouping for the samples or features included in
`data`. Expected to contain the following columns: `sample` - must match
samples in the `data` object, `group` - a grouping variable, e.g.,
condition or treatment, `replicate` - replicate name.}

\item{features}{A data frame that maps feature identifiers to names and
descriptions. Expected to contain the following columns: `id` - feature id,
must match the identifier in the `data` object, `name` - human-friendly
name of the feature, e.g. gene symbol, `description` - a brief description
of the feature.}

\item{fterms}{An object containing functional term information, created using
function \code{download_functional_terms}.}
}
\value{
The function does not return a value but launches a Shiny application
  in the user's default web browser, allowing for interactive exploration of
  the differential expression data.
}
\description{
This function creates and launches a Shiny application designed for exploring
differential expression (DE) data. It integrates various data inputs and
initializes interactive visualization modules for an enhanced data
exploration experience. The application offers a sidebar layout with themed
UI components and a range of interactive modules including global input,
volume-magnitude plot, feature plot, feature information, and enrichment
analysis.
}
\examples{
if(interactive()) {
  data(de, data, metadata, features)
  fterms <- download_functional_terms("yeast", feature_name = "gene_id")
  run_app(de, data, metadata, features, fterms)
}
}
